+++ src/types.ts
@@ -53,8 +53,6 @@ export interface PostHogConfig {
     ui_host: string | null
     token: string
     autocapture: boolean | AutocaptureConfig
-    // this is currently in development and may have breaking changes without a major version bump
-    autocapture_exceptions: boolean
     rageclick: boolean
     cross_subdomain_cookie: boolean
     persistence: 'localStorage' | 'cookie' | 'memory' | 'localStorage+cookie' | 'sessionStorage'